[WIP] Copy of #30174 Fix suggest_memory_format() for ambiguous cases
This PR is to fix the ambiguity in NC11 and N1HW tensor when specifying
memory_format.
Previous implementation unnecessarily associate memory contiguity with memory
format, which undermines the expressiveness of memory_format for the ambiguous
cases.

e.g. Previously, if we have NC11 tensor in channels_last format. Calling
contiguous on that tensor doesn't change the memory_format to NCHW.

ghstack-source-id: 793c89f4608d1f23b7daee435ff7b117beafbf20
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30325