[quant][graphmode] Supporting making submodules unique
Summary:
Support quantization for modules with reused submodules, e.g. relu (automatically make unique)
We first do a pass on the graph to find all duplicate uses of the same module, and record the `Value`s of the
module instance, for each of these values we create a new module and change the access to that module.

Test Plan:
python test/test_jit.py
Reviewers:
mvz

Subscribers:

Tasks:

Tags:

ghstack-source-id: c236bf58abe1318ece3034d494823b56fb517260
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30037